refactor: drop effects class overhead
